+++ src/instructlab/configuration.py
@@ -4,6 +4,7 @@
 from os import path
 from re import match
 from typing import Any, Optional, Union
+import enum
 import logging
 import os
 import sys
@@ -14,6 +15,7 @@ import typing
 # pylint: disable=ungrouped-imports
 from instructlab.training import (
     DeepSpeedOptions,
+    DistributedBackend,
     FSDPOptions,
     LoraOptions,
     TorchrunArgs,
@@ -375,7 +377,11 @@ class _train(BaseModel):
     """Class describing configuration of the 'train' sub-command."""
 
     # model configuration
-    model_config = ConfigDict(extra="ignore", protected_namespaces=())
+    model_config = ConfigDict(
+        extra="ignore",
+        protected_namespaces=(),
+        use_enum_values=True,  # populate models with the value property of enums, rather than the raw enum.
+    )
     model_path: str = Field(
         default=DEFAULTS.MODEL_REPO,
         description="Directory where the model to be trained is stored.",
@@ -419,6 +425,11 @@ class _train(BaseModel):
     )
     fsdp_cpu_offload_optimizer: bool = Field(
         default=False, description="Allow CPU offload for FSDP optimizer."
+    )
+    distributed_training_backend: DistributedBackend = Field(
+        default=DistributedBackend.DEEPSPEED,
+        description="Pick a distributed training backend framework for GPU accelerated full fine-tuning.",
+        validate_default=True,  # ensures that the 'use_enum_values' flag takes effect on the default value
     )
     lora_rank: int | None = Field(
         default=4,
@@ -956,6 +967,9 @@ def config_to_commented_map(
             if default_value is PydanticUndefined:
                 if default_factory is not None and callable(default_factory):
                     default_value = default_factory()
+            # When the default value's type is an Enum or Enum value, convert it to a string
+            elif isinstance(default_value, enum.Enum):
+                default_value = default_value.value
             set_comment(
                 cm, field_name, description, default_value, deprecated, examples, indent
             )
@@ -1287,6 +1301,9 @@ def map_train_to_library(ctx, params):
 
     train_args.deepspeed_options = ds_args
     train_args.fsdp_options = fsdp_args
+    train_args.distributed_backend = DistributedBackend(
+        params["distributed_training_backend"]
+    )
     train_args.lora = lora_args
     if params["pipeline"] == "full":
         train_args.disable_flash_attn = True